+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Suppression comments in `@examples` and `@examplesIf` now work correctly (#443).
 
+* Fix alignment of diagnostic in console when source uses tabs (#445).
+
 ## 0.5.0
 
 ### Deprecations